Ver1.6.0 サンプルコードの更新

+++ Cocos2d-x_v3/samples/basic/Classes/SSPlayer/SS6PlayerPlatform.cpp
@@ -583,7 +583,7 @@ namespace ss
 
 		MultiplyMatrix(pls.mat, mat, mat);
 
-		for (int i = 0; i < sprite->_meshVertexSize; i++)
+		for (size_t i = 0; i < sprite->_meshVertexSize; i++)
 		{
 			sprite->_mesh_colors[i * 4 + 0] = setcol[0];
 			sprite->_mesh_colors[i * 4 + 1] = setcol[1];
